extracted entry point test from previous class, extracted deserializer
