+++ src/test/java/com/mindlinksoft/recruitment/mychat/CommandLineArgumentParserTests.java
@@ -1,0 +1,14 @@
+package com.mindlinksoft.recruitment.mychat;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+public class CommandLineArgumentParserTests {
+
+	@Test
+	public void test() {
+		fail("Not yet implemented");
+	}
+
+}
