+++ src/test/java/com/mindlinksoft/recruitment/mychat/ConversationExporterTests.java
@@ -134,74 +134,4 @@ public class ConversationExporterTests {
 		assertEquals("YES! I'm the head pie eater there...", ms[6].content);
 	}
 
-	/**command line argument test*/
-	@Test
-	public void testCLI() throws IOException {
-		MainCLI.main(new String[] { "chat.txt", "chat.json" });
-
-		GsonBuilder builder = new GsonBuilder();
-		builder.registerTypeAdapter(Instant.class, new InstantDeserializer());
-
-		Gson g = builder.create();
-
-		Conversation c = g.fromJson(new InputStreamReader(new FileInputStream("chat.json")), Conversation.class);
-		Files.delete(FileSystems.getDefault().getPath("chat.json"));
-		assertNotNull(c);
-		assertEquals("My Conversation", c.name);
-
-		assertEquals(7, c.messages.size());
-
-		Message[] ms = new Message[c.messages.size()];
-		c.messages.toArray(ms);
-
-		assertEquals(Instant.ofEpochSecond(1448470901), ms[0].timestamp);
-		assertEquals("bob",ms[0].senderId);
-		assertEquals("Hello there!", ms[0].content);
-
-		assertEquals(Instant.ofEpochSecond(1448470905), ms[1].timestamp);
-		assertEquals("mike", ms[1].senderId);
-		assertEquals("how are you?", ms[1].content);
-
-		assertEquals(Instant.ofEpochSecond(1448470906), ms[2].timestamp);
-		assertEquals("bob", ms[2].senderId);
-		assertEquals("I'm good thanks, do you like pie?", ms[2].content);
-
-		assertEquals(Instant.ofEpochSecond(1448470910), ms[3].timestamp);
-		assertEquals("mike", ms[3].senderId);
-		assertEquals("no, let me ask Angus...", ms[3].content);
-
-		assertEquals(Instant.ofEpochSecond(1448470912), ms[4].timestamp);
-		assertEquals("angus", ms[4].senderId);
-		assertEquals("Hell yes! Are we buying some pie?", ms[4].content);
-
-		assertEquals(Instant.ofEpochSecond(1448470914), ms[5].timestamp);
-		assertEquals("bob", ms[5].senderId);
-		assertEquals("No, just want to know if there's anybody else in the pie society...", ms[5].content);
-
-		assertEquals(Instant.ofEpochSecond(1448470915), ms[6].timestamp);
-		assertEquals("angus", ms[6].senderId);
-		assertEquals("YES! I'm the head pie eater there...", ms[6].content);
-	}
-
-	private void verifyPieConversation(Conversation c) {
-
-	}
-
-	class InstantDeserializer implements JsonDeserializer<Instant> {
-
-		@Override
-		public Instant deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
-			if (!jsonElement.isJsonPrimitive()) {
-				throw new JsonParseException("Expected instant represented as JSON number, but no primitive found.");
-			}
-
-			JsonPrimitive jsonPrimitive = jsonElement.getAsJsonPrimitive();
-
-			if (!jsonPrimitive.isNumber()) {
-				throw new JsonParseException("Expected instant represented as JSON number, but different primitive found.");
-			}
-
-			return Instant.ofEpochSecond(jsonPrimitive.getAsLong());
-		}
-	}
 }
